pyproject.toml
[project]
name = "music-memes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "lxml>=6.0.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "regex>=2025.7.34",
    "rich>=14.1.0",
    "scikit-image>=0.25.2",
    "spotipy>=2.25.1",
    "streamlit>=1.48.0",
    "typer>=0.16.0",
]
